+++ src/openrct2/network/NetworkServerAdvertiser.cpp
@@ -13,6 +13,7 @@
 
 #    include "../config/Config.h"
 #    include "../core/Console.hpp"
+#    include "../core/Json.hpp"
 #    include "../core/String.hpp"
 #    include "../core/Util.hpp"
 #    include "../localisation/Date.h"
